extract text highlight creation into directive
