+++ projects/media-viewer/src/lib/media-viewer.component.html
@@ -48,7 +48,8 @@
                    [downloadFileName]="downloadFileName"
                    [height]="height"
                    [caseId]="caseId"
-                   rotationPersist>
+                   rotationPersist
+                   mvCreateTextHighlight>
     </mv-pdf-viewer>
     <mv-image-viewer *ngIf="contentType === 'image'"
                      (mediaLoadStatus)="onMediaLoad($event)"
