+++ projects/media-viewer/src/lib/redaction/components/redaction.component.ts
@@ -62,7 +62,7 @@ export class RedactionComponent implements OnInit, OnDestroy {
   }
 
   markTextRedaction(highlight) {
-    const redactionHighlight = this.highlightService.getRectangles(highlight.event, highlight.page);
+    const redactionHighlight = highlight.rectangles;
     if (redactionHighlight && redactionHighlight.length) {
       this.saveRedaction(highlight.page, [...redactionHighlight])
     }
